website/backend/requirements_py312.txt
fastapi
uvicorn
sentence-transformers
pydantic-settings
google-generativeai
langchain
langchain-community
langchain-core
langchain-text-splitters
qdrant-client
cohere
python-dotenv
numpy
pytest
python-jose[cryptography]
passlib[bcrypt]
googletrans==3.1.0a0  # Use an alpha version that's compatible with newer httpx